MiscClients/golang_rest_service/python_client/requirements.txt
requests==2.32.4
PyJWT==2.10.1
kafka-python==2.2.15
